Fix database session leak in WebSocket use of auth handlers
WebSocket connections stay open longer which was causing DB sessions
in auth handlers to leak. (HTTP requests were fine because the
autocommit approach was cleaning things up correctly). Now, added a
context manager to explicitly commit and close DB sessions. Some is
overkill as the autocommit for HTTP requests is fine but adding the
context manager anyway just to be consistent and in case they're
ever used in WebSocket routes.