task/WG-594 improve db connections (#260)
* Add app name to db connection for debuggin/monitoring

* Fix database connection pool exhaustion from Celery tasks

- Create ONE shared engine/pool for all Celery tasks via get_celery_engine()
- Add application_name to connection strings for monitoring
- Rename engine to migrate_engine for clarity

* Fix formatting

* Fix error when renaming enging to migrate_engine

* Reconfigure connections for Litestar

* Use env in db connections' application_name

* Remove unused method

Added in https://github.com/TACC-Cloud/geoapi/pull/248 but not used there.

* Rename litestar engine_config

* Fix use of APP_ENV

* Fix mock

* Fix formatting

* Remove line change

* Update comment and celery max_overflow

* Fix database session leak in WebSocket use of auth handlers

WebSocket connections stay open longer which was causing DB sessions
in auth handlers to leak. (HTTP requests were fine because the
autocommit approach was cleaning things up correctly). Now, added a
context manager to explicitly commit and close DB sessions. Some is
overkill as the autocommit for HTTP requests is fine but adding the
context manager anyway just to be consistent and in case they're
ever used in WebSocket routes.

* Add shorter WebSocket timeout for local dev

* Add log statement

* Fix import

* Revert "Add shorter WebSocket timeout for local dev"

This reverts commit c7a5fdf43f45c169fc4971dae98cf86fd371ad11.